Backend Finished Starting on frontend integration with redux  previous react native code

+++ backend/controller/ingredientController.js
@@ -1,11 +1,13 @@
 const asyncHandler = require("express-async-handler")
 
 const Ingredient = require("../models/ingredientModel")
+const User = require("../models/userModel")
 //@desc     Get all user ingredients
 //@route     GET /api/ingredients
 //@access     Private
 const getIngrs = asyncHandler ( async (req, res) =>{
-    const ingredients = await Ingredient.find(); 
+    // find this user's specific ingredients
+    const ingredients = await Ingredient.find({user: req.user.id}); 
 
     res.status(200).json(ingredients) 
 })
@@ -20,7 +22,8 @@ const setIngr = asyncHandler ( async (req, res) =>{
     }
 
     const ingredient = await Ingredient.create({
-        name: req.body.name
+        name: req.body.name,
+        user:req.user.id, // attach this to current user 
     })
     res.status(200).json(ingredient)
 })
@@ -36,6 +39,21 @@ const updateIngr = asyncHandler ( async (req, res) =>{
         throw new Error("Ingredient Not Found")
     }
 
+    const user = await User.findById(req.user.id)
+    
+    // check for user 
+    if(!user){
+        res.status(401)
+        throw new Error("User Not Found")
+    }
+
+    // make sure the logged in user matches the ingredient user 
+    if(ingredient.user.toString() !== user.id){
+        res.status(401)
+        throw new Error("User not authorized")
+    }
+    
+    
     const updatedIngredient = await Ingredient.findByIdAndUpdate(req.params.id, req.body, {
         new:true
     })
@@ -53,6 +71,22 @@ const deleteIngr = asyncHandler ( async (req, res) =>{
         throw new Error("Ingredient Not Found")
     }
 
+
+    const user = await User.findById(req.user.id)
+    
+    // check for user 
+    if(!user){
+        res.status(401)
+        throw new Error("User Not Found")
+    }
+
+    // make sure the logged in user matches the ingredient user 
+    if(ingredient.user.toString() !== user.id){
+        res.status(401)
+        throw new Error("User not authorized")
+    }
+    
+    
     await ingredient.deleteOne() 
 
     res.status(200).json({id: req.params.id})
